feat: issue-9: remove types.ts file from organisms/UiPhoneNumber directory and export types directly from the UiPhoneNumber.vue
