--- src/components/organisms/UiPhoneNumber/types.ts
@@ -1,10 +1,0 @@
-export interface UiPhoneNumberProps {
-  label: string,
-  placeholder: string,
-  errorMessage: string,
-  modelValue: string,
-  touched: boolean,
-  language: string,
-  defaultCountryCode: string,
-  hint: string | boolean,
-}
